rename group hook back to segments
